# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tenders": {
            "get": {
                "tags": [
                    "tender"
                ],
                "summary": "Возращает тендеры",
                "description": "возращает тендеры с возможностью фильрации",
                "operationId": "tenders",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "number",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "source_url",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_requset_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "timestamp"
                        }
                    },
                    {
                        "name": "end_requset_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "timestamp"
                        }
                    },
                    {
                        "name": "result_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "timestamp"
                        }
                    },
                    {
                        "name": "nmc_price",
                        "in": "query",
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "ensure_request_price",
                        "in": "query",
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "ensure_contract_price",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "type_id",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "currency_id",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "stage_id",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "okvad2_classifier",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "customer_name",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "customer_inn",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "customer_ogrn",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "customer_kpp",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "customer_location",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "customer_contact_phone",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "customer_contact_name",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type_name",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "currency",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "currency_name",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "stage_name",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "objects_name",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "objects_okvad",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Тендеры не найдены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Тендеров соответствующих таким фильтрам нет"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Пользователь не зарегистрирован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Пользователь не зарегистрирован"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Возвращенные тендеры",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Tender": {
                                            "$ref": "#/components/schemas/Tender"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Tender": {
                "title": "Tender model",
                "description": "Tender model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "default": 1
                    },
                    "number": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "default": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string",
                        "format": "placed",
                        "default": "Ms. Natasha Effertz PhD"
                    },
                    "description": {
                        "title": "description",
                        "description": "description",
                        "type": "string",
                        "format": "placed",
                        "default": "Non id et similique vel omnis quidem quaerat."
                    },
                    "source_url": {
                        "title": "source_url",
                        "description": "source_url",
                        "type": "string",
                        "format": "placed",
                        "default": "https://www.lockman.info/dolorum-aperiam-autem-quos-ipsam-aut"
                    },
                    "start_request_date": {
                        "title": "start_request_date",
                        "description": "start_request_date",
                        "type": "string",
                        "format": "timestamp",
                        "default": "2014-03-07 05:45:31"
                    },
                    "end_request_date": {
                        "title": "end_request_date",
                        "description": "end_request_date",
                        "type": "string",
                        "format": "timestamp",
                        "default": "2022-04-08 02:34:31"
                    },
                    "result_date": {
                        "title": "result_date",
                        "description": "result_date",
                        "type": "string",
                        "format": "timestamp",
                        "default": "2023-01-05 23:12:07"
                    },
                    "nmc_price": {
                        "title": "nmc_price",
                        "description": "nmc_price",
                        "type": "string",
                        "format": "placed",
                        "default": "3459.86"
                    },
                    "ensure_request_price": {
                        "title": "ensure_request_price",
                        "description": "ensure_request_price",
                        "type": "string",
                        "format": "placed",
                        "default": "9640.54"
                    },
                    "ensure_contract_price": {
                        "title": "ensure_contract_price",
                        "description": "ensure_contract_price",
                        "type": "string",
                        "format": "placed",
                        "default": "3500.85"
                    },
                    "customer_id": {
                        "title": "customer_id",
                        "description": "customer_id",
                        "type": "integer",
                        "format": "int32",
                        "default": 11
                    },
                    "type_id": {
                        "title": "type_id",
                        "description": "type_id",
                        "type": "integer",
                        "format": "int32",
                        "default": 1
                    },
                    "currency_id": {
                        "title": "currency_id",
                        "description": "currency_id",
                        "type": "integer",
                        "format": "int32",
                        "default": 11
                    },
                    "stage_id": {
                        "title": "stage_id",
                        "description": "stage_id",
                        "type": "integer",
                        "format": "int32",
                        "default": 1
                    }
                },
                "type": "object"
            },
...
        }
    },
    "security": [
        []
    ]
}